Extract package managers into their own files
